pyproject.toml
[project]
name = "custom-user-verifier"
version = "0.1.0"
description = "Custom User Verifier Flask server for Arcade.dev"
requires-python = ">=3.10"
dependencies = [
    "flask",
    "arcadepy",
    "python-dotenv",
]

[project.scripts]
verifier = "app:main"
